+++ src/utils/testUtils.ts
@@ -26,6 +26,7 @@ export const mockLogger: FastifyLoggerInstance = {
 }
 
 export const getMockDataStorageClient = (props: Partial<DataStorageClient> = {}): DataStorageClient => ({
+  getAllInteractions: props.getAllInteractions || jest.fn(),
   createInteraction: props.createInteraction || jest.fn(),
   abortInteraction: props.abortInteraction || jest.fn(),
   getOngoingInteractions: props.getOngoingInteractions || jest.fn(),